Use a faster way to load reminders
Instead of loading them one by one, load all at once and then associate them with the correct item.
This seems to avoid quite a lot of overhead and speeds the loading process up for a reasonable amount of reminders.
The algorithm to do the association is probably not ideal yet, since it does linear searches for O(n^2) performance.