Complete reimagining of the notification handling
Instead of giving the android side access to the db, have everything go through channels
back to the flutter/dart side.
This means that all database operations are now contained in flutter.
The disadvantage of this approach is the introduction of a background flutter engine, that
is used when the app itself is not open.