Switch to Navigator 2.0 and introduce the base bloc
The navigation method, in particular for the reaction to a tapped notification, was not ideal before.
The main issue was that the app didn't really have a way to track which state it's in and that therefore
reacting to a notification was quite a hack. In particular, when the currently selected list and the list
of the tapped item are the same.

The new Navigator 2.0 method in Flutter is much more verbose than the older version, but allows to properly
modify the navigation state of the app.
The state is held in a new base bloc, which deals with transitions between the states.